types: consolidate shared contract types into src/shared/types.ts (bpdiscord-z57)
Move the 9 types duplicated between src/server/types.ts and src/client/types.ts
(ApiResponse, User, AuthRequest, SignupRequest, AuthResponse,
PasswordResetRequest, PasswordResetConfirmRequest, ScraperRequest,
AuthenticatedUser) into src/shared/types.ts — already wired into both tsconfigs
via ../shared/**/*. Side-specific types (Express-, React-, Drizzle-, Puppeteer-
coupled) stay where they are.

Reconciled two divergent shapes during the move:
- ApiResponse.details: server used any, client used unknown[]. Picked unknown
  (server comment confirmed it's used as array/object/string).
- AuthResponse.user: server used any, client used AuthenticatedUser. Picked
  AuthenticatedUser and loosened email to optional to match Supabase's actual
  shape; re-tightened to required in StoredUser, where isValidStoredUser
  already enforces it at runtime.

Also deleted 20 unused exports surfaced by the audit:
- server: RatingData, UserRatingsData, LBFilm
- client: Rating (locally redeclared in 3 components), FilmUserApiResponse,
  UserRatingsResponse, HaterRankingsResponse, ComparisonResponse,
  UseAuthReturn, UseApiReturn, UseLocalStorageReturn, AppError,
  ErrorBoundaryState, FormState, LoginFormData, SignupFormData,
  RATING_VALUES, RatingValue, LoadingState, AuthStatus

Verification:
- Server tsc --noEmit: clean
- Client tsc --noEmit: 37 errors, all pre-existing on main (0 new from this PR)
- Client tests: 183/183 pass
- Server tests: 114/134 pass (20 pre-existing DB integration failures, baseline
  on main is 112/134 — this PR fixes 2 flakies, regresses 0)
