+++ src/client/components/LoginPage.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import PasswordReset from "./PasswordReset";
 import apiService from "../services/api";
-import { AuthRequest, SignupRequest } from "../types";
+import { AuthRequest, SignupRequest } from "../../shared/types";
 import { Subheading } from "./Subheading";
 
 const LoginPage = () => {
