+++ src/client/types.ts
@@ -1,72 +1,12 @@
-// ===========================
-// API Response Types
-// ===========================
+// Cross-cutting API/contract types (ApiResponse, User, AuthRequest,
+// SignupRequest, AuthResponse, PasswordResetRequest,
+// PasswordResetConfirmRequest, ScraperRequest, AuthenticatedUser) live in
+// src/shared/types.ts. Import from there directly.
 
-export interface ApiResponse<T = unknown> {
-  data?: T;
-  message?: string;
-  error?: string;
-  count?: number;
-  details?: unknown[];
-}
+import type { AuthenticatedUser } from "../shared/types";
 
 // ===========================
-// User & Authentication Types
-// ===========================
-
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface AuthRequest {
-  email: string;
-  password: string;
-}
-
-export interface SignupRequest extends AuthRequest {
-  name: string;
-}
-
-export interface AuthResponse {
-  message: string;
-  access_token?: string;
-  user: AuthenticatedUser;
-}
-
-// ===========================
-// Password Reset Types
-// ===========================
-
-export interface PasswordResetRequest {
-  email: string;
-}
-
-export interface PasswordResetConfirmRequest {
-  email?: string;
-  password: string;
-  token?: string;
-}
-
-// ===========================
-// Scraper Types (for API calls)
-// ===========================
-
-export interface ScraperRequest {
-  url: string;
-  selectors: Array<{
-    name: string;
-    css: string;
-    attributes?: string[];
-    multiple?: boolean;
-  }>;
-}
-
-// ===========================
-// UI Component Types
+// UI Data Shapes
 // ===========================
 
 export interface Rating {
@@ -99,22 +39,13 @@ export interface UserComparisonData {
 }
 
 // ===========================
-// Authentication & User Types (Enhanced)
+// Client-side Stored Auth
 // ===========================
 
-export interface AuthenticatedUser {
-  id: string;
-  email: string;
-  user_metadata?: {
-    name?: string;
-  };
-  app_metadata?: {
-    provider?: string;
-    providers?: string[];
-  };
-}
-
 export interface StoredUser extends AuthenticatedUser {
+  // Re-required: a stored (logged-in) user always has an email — the
+  // isValidStoredUser guard below enforces this at runtime.
+  email: string;
   aud: string;
   role?: string;
 }
@@ -132,7 +63,7 @@ export const isValidStoredUser = (obj: unknown): obj is StoredUser => {
 };
 
 // ===========================
-// Movie & Rating Types (Enhanced)
+// Movie Comparison
 // ===========================
 
 export interface MovieInCommon {
@@ -158,43 +89,6 @@ export interface TasteCompatibility {
   description: string;
   commonMovies: number;
   averageRatingDifference: number;
-}
-
-// ===========================
-// API Response Types (Specific)
-// ===========================
-
-export interface FilmUserApiResponse {
-  username: string;
-  displayName?: string;
-  followers?: number;
-  following?: number;
-  numberOfLists?: number;
-  totalRatings?: number;
-  ratings: Rating[];
-  source: "database" | "scraped" | "scraped_fallback" | "mixed_fallback";
-  success: boolean;
-}
-
-export interface UserRatingsResponse {
-  username: string;
-  ratings: Rating[];
-  totalRatings: number;
-  averageRating?: number;
-  source: string;
-}
-
-export interface HaterRankingsResponse {
-  rankings: HaterRanking[];
-  totalUsers: number;
-  lastUpdated?: string;
-}
-
-export interface ComparisonResponse {
-  user1: UserData;
-  user2: UserData;
-  moviesInCommon?: MoviesInCommonData;
-  tasteCompatibility?: TasteCompatibility;
 }
 
 // ===========================
@@ -251,72 +145,6 @@ export interface ProtectedRouteProps {
 export interface HaterRankingsProps {
   onBackToProfile?: () => void;
   isPublic?: boolean;
-}
-
-// ===========================
-// Hook Return Types
-// ===========================
-
-export interface UseAuthReturn {
-  user: StoredUser | null;
-  isAuthenticated: boolean;
-  isLoading: boolean;
-  login: (credentials: AuthRequest) => Promise<void>;
-  signup: (userData: SignupRequest) => Promise<void>;
-  logout: () => void;
-  error: string | null;
-}
-
-export interface UseApiReturn<T> {
-  data: T | null;
-  loading: boolean;
-  error: string | null;
-  execute: () => Promise<void>;
-  reset: () => void;
-}
-
-export interface UseLocalStorageReturn<T> {
-  value: T | null;
-  setValue: (value: T | null) => void;
-  removeValue: () => void;
-  error: string | null;
-}
-
-// ===========================
-// Error Types
-// ===========================
-
-export interface AppError {
-  message: string;
-  code?: string | number;
-  details?: unknown;
-  timestamp: Date;
-}
-
-export interface ErrorBoundaryState {
-  hasError: boolean;
-  error: Error | null;
-  errorInfo: React.ErrorInfo | null;
-}
-
-// ===========================
-// Form Types
-// ===========================
-
-export interface FormState {
-  isSubmitting: boolean;
-  errors: Record<string, string>;
-  touched: Record<string, boolean>;
-}
-
-export interface LoginFormData {
-  email: string;
-  password: string;
-}
-
-export interface SignupFormData extends LoginFormData {
-  name: string;
-  confirmPassword?: string;
 }
 
 // ===========================
@@ -474,22 +302,3 @@ export interface UserScrapeJob extends RefreshJob {
   lbusername: string;
 }
 
-// ===========================
-// Constants & Enums
-// ===========================
-
-export const RATING_VALUES = [0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5] as const;
-export type RatingValue = (typeof RATING_VALUES)[number];
-
-export enum LoadingState {
-  IDLE = "idle",
-  LOADING = "loading",
-  SUCCESS = "success",
-  ERROR = "error",
-}
-
-export enum AuthStatus {
-  AUTHENTICATED = "authenticated",
-  UNAUTHENTICATED = "unauthenticated",
-  LOADING = "loading",
-}
